# PERF: avoid rebuilding blknos